Completetly restructure compiler
