+++ Cargo.toml
@@ -12,6 +12,7 @@ llvm-sys = "140"
 regex = "1.9.1"
 tempfile = "3.5.0"
 libc = "0.2"
+rand = "0.8"
 
 [lib]
 path = "staticlib/builtins.rs"
