+++ src/compiler.rs
@@ -1,10 +1,6 @@
 extern crate inkwell;
 
 use std::collections::HashMap;
-use std::env::var;
-use std::error::Error;
-use std::fmt::Pointer;
-use std::hash::Hash;
 use std::path::Path;
 use std::process::Command;
 
@@ -13,359 +9,41 @@ use inkwell::context::Context;
 use inkwell::module::{Linkage, Module};
 use inkwell::support::LLVMString;
 use inkwell::targets::{
-    CodeModel, FileType, InitializationConfig, RelocMode, Target, TargetData, TargetMachine,
+    CodeModel, FileType, InitializationConfig, RelocMode, Target, TargetMachine,
 };
-use inkwell::types::{
-    AnyType, AnyTypeEnum, BasicMetadataTypeEnum, BasicType, BasicTypeEnum, PointerType,
-};
+use inkwell::types::{BasicMetadataTypeEnum, BasicType, BasicTypeEnum};
 use inkwell::values::{
-    AnyValue, AnyValueEnum, BasicMetadataValueEnum, BasicValue, BasicValueEnum, FunctionValue,
-    IntValue, PointerValue,
+    AnyValue, AnyValueEnum, BasicMetadataValueEnum, BasicValueEnum, FunctionValue, PointerValue,
 };
-use inkwell::{execution_engine, AddressSpace, OptimizationLevel};
+use inkwell::{AddressSpace, OptimizationLevel};
 
 use crate::ast::{ASTNode, Argument, Operator, VariableType};
+use crate::common_utils::into_basic_value_enum;
+use crate::function::PyroFunction;
 
-/// Turn an AnyValueEnum into a BasicValueEnum if possible.
-///
-/// If there is no relevant BasicValueEnum for the given AnyValueEnum,
-/// an Err is returned.
-///
-/// # Note
-/// At the current moment, this functions is not implemented for all
-/// possible BasicValueEnums.
-fn into_basic_value_enum(value: AnyValueEnum) -> Result<BasicValueEnum, String> {
-    match value.get_type() {
-        AnyTypeEnum::IntType(_) => Ok(value.into_int_value().as_basic_value_enum()),
-        AnyTypeEnum::FloatType(_) => Ok(value.into_float_value().as_basic_value_enum()),
-        AnyTypeEnum::ArrayType(_) => Ok(value.into_array_value().as_basic_value_enum()),
-        AnyTypeEnum::PointerType(_) => Ok(value.into_pointer_value().as_basic_value_enum()),
-        _ => Err(format!(
-            "No basic value could be constructed from {:?}",
-            value
-        )),
-    }
-}
-
-struct FunctionCompiler<'a, 'ctx> {
+/// Compiler for the Pyro programming language.
+pub struct Compiler<'ctx> {
     context: &'ctx Context,
-    module: &'a Module<'ctx>,
-    builder: &'a Builder<'ctx>,
-
-    identifier: String,
-    arguments: Vec<Argument>,
-    body: Vec<ASTNode>,
-    return_type: VariableType,
+    module: Module<'ctx>,
+    builder: Builder<'ctx>,
 
     constant_idx: u64,
-    variables: HashMap<String, PointerValue<'ctx>>,
     string_constants: HashMap<String, PointerValue<'ctx>>,
-    fn_value: Option<FunctionValue<'ctx>>,
 }
-impl<'a, 'ctx> FunctionCompiler<'a, 'ctx> {
-    fn compile_function(
-        context: &'ctx Context,
-        module: &'a Module<'ctx>,
-        builder: &'a Builder<'ctx>,
-        function_declaration: ASTNode,
-    ) -> Result<FunctionValue<'ctx>, String> {
-        let mut function;
 
-        if let ASTNode::FunctionDeclaration {
-            identifier,
-            arguments,
-            return_type,
-            body,
-        } = function_declaration
-        {
-            function = FunctionCompiler {
-                context,
-                module,
-                builder,
-
-                identifier,
-                arguments,
-                body,
-                return_type,
-
-                constant_idx: 0,
-                variables: HashMap::new(),
-                string_constants: HashMap::new(),
-                fn_value: None,
-            }
-        } else {
-            return Err(format!(
-                "Not a valid Function Declaration {:?}",
-                function_declaration
-            ));
-        }
-        function.compile()
-    }
-
-    /// Compile the interface defined by a FunctionDeclaration.
-    fn get_function_prototype(&mut self) -> Result<FunctionValue<'ctx>, String> {
-        let ret_type = self.get_type_from_variable_type(&self.return_type);
-
-        let i64_type = self.context.i64_type(); // TODO: As above
-
-        let args_types = self
-            .arguments
-            .iter()
-            .map(|a| {
-                self.get_type_from_variable_type(&a.argument_type)
-                    .expect("Invalid Argument type")
-                    .into()
-            })
-            .collect::<Vec<BasicMetadataTypeEnum>>();
-        let args_types = args_types.as_slice();
-
-        let fn_type = match ret_type {
-            Some(ret_type) => ret_type.fn_type(args_types, false),
-            None => self.context.void_type().fn_type(args_types, false),
-        };
-        let fn_val = self
-            .module
-            .add_function(self.identifier.as_str(), fn_type, None);
-        self.fn_value = Some(fn_val);
-
-        for (i, arg) in fn_val.get_param_iter().enumerate() {
-            arg.set_name(self.arguments[i].identifier.as_str());
-        }
-        Ok(fn_val)
-    }
-
-    fn compile(&mut self) -> Result<FunctionValue<'ctx>, String> {
-        let function = self.get_function_prototype()?;
-        let entry = self.context.append_basic_block(function, "entry");
-        self.fn_value = Some(function);
-
-        self.variables.reserve(self.arguments.len());
-        self.builder.position_at_end(entry);
-
-        for (i, arg) in function.get_param_iter().enumerate() {
-            let argument_name = self.arguments[i].clone().identifier;
-            let alloced = self.allocate_entry_stack_block(argument_name.as_str(), arg.get_type());
-            self.builder.build_store(alloced, arg);
-
-            self.variables.insert(argument_name, alloced);
-        }
-
-        let body = self.build_function_body();
-
-        if self.return_type == VariableType::Void {
-            self.builder.build_return(None);
-        }
-        Ok(function)
-    }
-
-    /// Allocate a block of memory on the stack in the function entry block .
-    /// NOTE: Currently only allocates memory of size i64.
-    fn allocate_entry_stack_block<T: BasicType<'ctx>>(
-        &mut self,
-        name: &str,
-        ty: T,
-    ) -> PointerValue<'ctx> {
-        let local_builder = self.context.create_builder();
-        let entry = self.fn_value.unwrap().get_first_basic_block().unwrap();
-
-        match entry.get_first_instruction() {
-            Some(first_instr) => local_builder.position_before(&first_instr),
-            None => local_builder.position_at_end(entry),
-        }
-        local_builder.build_alloca(ty, name)
-    }
-
-    fn get_type_from_variable_type(
-        &self,
-        variable_type: &VariableType,
-    ) -> Option<BasicTypeEnum<'ctx>> {
-        match variable_type {
-            VariableType::Integer => Some(self.context.i64_type().as_basic_type_enum()),
-            VariableType::Void => None,
-            _ => unimplemented!(),
-        }
-    }
-
-    /// Build a binary operation out of the given operands.
+impl<'ctx> Compiler<'ctx> {
+    /// Create a new compiler using the provided Inkwell LLVM context.
     ///
-    /// The `left` operand will determine the type of the operation.
-    /// The `right` operand needs to be of the same type for operation to work.
-    ///
-    /// Returns the resulting LLVM MathValue or an error if the types did not match.
-    fn build_binary_operation(
-        &mut self,
-        left: BasicValueEnum<'ctx>,
-        right: BasicValueEnum<'ctx>,
-        operator: Operator,
-    ) -> Result<AnyValueEnum<'ctx>, String> {
-        let lhs;
-        let rhs;
-        let op;
+    /// The LLVM module used throughout compilation will use the name `module_name`.
+    pub fn new(module_name: String, context: &'ctx Context) -> Self {
+        Self {
+            context,
+            module: context.create_module(&module_name),
+            builder: context.create_builder(),
 
-        let op_type = left.get_type();
-
-        match op_type {
-            BasicTypeEnum::IntType(_) => match right.get_type() {
-                BasicTypeEnum::IntType(_) => {
-                    lhs = left.into_int_value();
-                    rhs = right.into_int_value();
-
-                    match operator {
-                        Operator::Plus => op = self.builder.build_int_add(lhs, rhs, "intaddition"),
-                        Operator::Minus => {
-                            op = self.builder.build_int_sub(lhs, rhs, "intsubtraction")
-                        }
-                        Operator::Multiplication => {
-                            op = self.builder.build_int_mul(lhs, rhs, "intmultiplication")
-                        }
-                        Operator::Division => {
-                            op = self.builder.build_int_signed_div(lhs, rhs, "intdivision")
-                        }
-                    }
-                }
-                _ => {
-                    return Err(format!(
-                        "{:?} and {:?} are not of the same type for add",
-                        left, right
-                    ))
-                }
-            },
-            _ => unimplemented!(),
+            constant_idx: 0,
+            string_constants: HashMap::new(),
         }
-
-        return Ok(op.as_any_value_enum());
-    }
-
-    fn get_next_constant_name(&mut self) -> String {
-        let mut string = String::new();
-        string.push_str("__pyro_compiler_constant_");
-        string.push_str(self.constant_idx.to_string().as_str());
-        self.constant_idx += 1;
-
-        string
-    }
-
-    /// Evaluate an expression from the AST and build the LLVM code for it.
-    /// Returns the final output type of the expression, or `Err` if no expression
-    /// could be parsed.
-    // TODO: This should be split up in separate methods to be less of an eyesore.
-    fn evaluate_statement(&mut self, statement: ASTNode) -> Result<AnyValueEnum<'ctx>, String> {
-        match statement {
-            ASTNode::IntegerLiteral(value) => {
-                let const_value = self.context.i64_type().const_int(value as u64, false);
-                Ok(const_value.as_any_value_enum())
-            }
-
-            ASTNode::StringLiteral(string) => {
-                if self.string_constants.contains_key(string.as_str()) {
-                    let ptr = self.string_constants.get(string.as_str()).unwrap();
-                    return Ok(ptr.as_any_value_enum());
-                }
-
-                let const_name = self.get_next_constant_name();
-                let ptr = self
-                    .builder
-                    .build_global_string_ptr(&string, const_name.as_str());
-                let ptr = ptr.as_pointer_value();
-                self.string_constants.insert(string, ptr);
-                Ok(ptr.as_any_value_enum())
-            }
-
-            ASTNode::Identifier(variable_name) => {
-                let variable_ptr = self.variables.get(&variable_name);
-                if variable_ptr.is_none() {
-                    return Err(format!(
-                        "Variable {} can not be found in scope.",
-                        variable_name
-                    ));
-                }
-                let variable_ptr = *variable_ptr.unwrap();
-                let variable = self.builder.build_load(variable_ptr, &variable_name);
-                Ok(variable.as_any_value_enum())
-            }
-
-            ASTNode::LetDeclaration(variable_name, expression) => {
-                let value = self.evaluate_statement(*expression)?;
-                let value = into_basic_value_enum(value)?;
-
-                let variable =
-                    self.allocate_entry_stack_block(variable_name.as_str(), value.get_type());
-                self.variables.insert(variable_name, variable);
-
-                self.builder.build_store(variable, value);
-                Ok(variable.as_any_value_enum())
-            }
-
-            ASTNode::BinaryOp {
-                operator,
-                left,
-                right,
-            } => {
-                let left = self.evaluate_statement(*left)?;
-                let right = self.evaluate_statement(*right)?;
-
-                let lhs = into_basic_value_enum(left)?;
-                let rhs = into_basic_value_enum(right)?;
-
-                self.build_binary_operation(lhs, rhs, operator)
-            }
-
-            ASTNode::FunctionCall(function_name, arguments) => {
-                let function = self.module.get_function(function_name.as_str()).unwrap();
-
-                let arguments = arguments
-                    .iter()
-                    .map(|a| self.evaluate_statement(a.clone()))
-                    .collect::<Result<Vec<AnyValueEnum>, String>>()?;
-
-                let arguments = arguments
-                    .iter()
-                    .map(|a| BasicMetadataValueEnum::try_from(*a))
-                    .collect::<Result<Vec<BasicMetadataValueEnum>, _>>();
-
-                if arguments.is_err() {
-                    return Err("Error when parsing arguments for function call".to_string());
-                }
-
-                let arguments = arguments.unwrap();
-                let call = self
-                    .builder
-                    .build_call(function, arguments.as_slice(), "call");
-
-                Ok(call.as_any_value_enum())
-            }
-
-            ASTNode::ReturnStatement(expression) => {
-                let value = self.evaluate_statement(*expression)?;
-                let value = into_basic_value_enum(value)?;
-
-                let return_instruction = self.builder.build_return(Some(&value));
-                Ok(return_instruction.as_any_value_enum())
-            }
-            _ => Err("Not implemented".to_string()),
-        }
-    }
-
-    fn build_function_body(&mut self) -> Result<(), String> {
-        let iterator = self.body.to_vec();
-        for statement in iterator {
-            let value = self.evaluate_statement(statement);
-            if value.is_err() {
-                return Err(value.unwrap_err());
-            }
-        }
-        Ok(())
-    }
-}
-
-pub struct Compiler {
-    context: Context,
-}
-
-impl<'ctx> Compiler {
-    pub fn new() -> Self {
-        let context = Context::create();
-        Self { context }
     }
 
     fn get_default_target_machine(&self) -> Result<TargetMachine, String> {
@@ -393,15 +71,16 @@ impl<'ctx> Compiler {
         Ok(target_machine.unwrap())
     }
 
-    fn add_print(&'ctx self, module: &Module<'ctx>) {
+    fn add_print(&self) {
         let void_type = self.context.void_type();
         let char_ptr = self.context.i8_type().ptr_type(AddressSpace::default());
         let print_type = void_type.fn_type(&[char_ptr.into()], true);
-        module.add_function("print", print_type, Some(Linkage::External));
+        self.module
+            .add_function("print", print_type, Some(Linkage::External));
     }
 
     pub fn compile(
-        &'ctx mut self,
+        mut self,
         ast: ASTNode,
         output_path: Option<&Path>,
     ) -> Result<Option<LLVMString>, String> {
@@ -409,23 +88,22 @@ impl<'ctx> Compiler {
             .get_default_target_machine()
             .expect("Error when creating target machine");
 
-        let module = self.context.create_module("name");
-        let builder = self.context.create_builder();
-        self.add_print(&module);
+        self.add_print();
 
         match ast {
             ASTNode::Program(functions) => {
                 for function_declaration in functions {
-                    FunctionCompiler::compile_function(
-                        &self.context,
-                        &module,
-                        &builder,
+                    PyroFunction::compile_function(
+                        self.context,
+                        &self.module,
+                        &self.builder,
+                        &mut self.string_constants,
                         function_declaration,
                     )?;
                 }
 
-                println!("{:?}", module.verify());
-                module.print_to_stderr();
+                println!("{:?}", self.module.verify());
+                self.module.print_to_stderr();
 
                 // TODO: Everything below this should be moved around
                 //       create functions and remove debug statements.
@@ -433,17 +111,17 @@ impl<'ctx> Compiler {
                 if output_path.is_some() {
                     if let Ok(tempfile) = tempfile::NamedTempFile::new() {
                         target_machine
-                            .write_to_file(&module, FileType::Object, tempfile.path())
+                            .write_to_file(&self.module, FileType::Object, tempfile.path())
                             .expect("Could not write module to file");
 
-                        let temppath = tempfile
-                            .path()
-                            .to_str()
-                            .expect("Error creating temporary file");
+                        // let temppath = tempfile
+                        //     .path()
+                        //     .to_str()
+                        //     .expect("Error creating temporary file");
                         let output_path = output_path.unwrap().to_str().unwrap();
 
                         // TODO: Make this actually look for library instead of hard coded debug path
-                        let out_str = Command::new("clang")
+                        Command::new("clang")
                             .args([
                                 "-no-pie",
                                 tempfile.path().to_str().unwrap(),
@@ -456,7 +134,7 @@ impl<'ctx> Compiler {
                     }
                     return Ok(None);
                 } else {
-                    return Ok(Some(module.print_to_string()));
+                    return Ok(Some(self.module.print_to_string()));
                 }
             }
             _ => unimplemented!(),
